Separated App.js into components
